fix: move appearance settings under file budget
